Quickbook: Deal with checked paths that aren't present. Refs #6691.
I think that the dependencies should just be files that were loaded, so
I changed it to do that, but added an extra (hidden for now) option to
list all paths that were checked, staring with a `+` for files that are
present, and `-` for files that are absent.

Also fixed a bug with the path resolver.

git-svn-id: https://svn.boost.org/svn/boost/trunk/tools/quickbook@77973 b8fc166d-592f-0410-95f2-cb63ce0dd405