Simpler, less efficient, id generation.
I noticed a bug for an edge case in the id generator. Wouldn't be too
hard to fix, but the implementation was too complicated with some really
pointless optimizations, so I rewrote it using a simpler brute force
method. Will be fine for now unless something has a lot of duplicate ids.

git-svn-id: https://svn.boost.org/svn/boost/trunk/tools/quickbook@85368 b8fc166d-592f-0410-95f2-cb63ce0dd405